requirements.txt
CacheControl==0.14.3
Faker==37.4.0
Jinja2==3.1.6
Mako==1.3.10
Markdown==3.8.2
MarkupSafe==3.0.2
PySocks==1.7.1
PyYAML==6.0.2
Pygments==2.19.2
Send2Trash==1.8.3
Werkzeug==3.1.3
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
alembic==1.15.2
analyzer==0.1.1
analyzerdam==0.1.2
analyzerstrategies==0.1.5
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi-bindings==21.2.0
argon2-cffi==25.1.0
ariadne==0.26.2
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
auth==0.5.3
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
boolean.py==5.0
botocore==1.38.23
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
cryptography==44.0.1
cyclonedx-python-lib==9.1.0
databases==0.9.0
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
eventlet==0.40.0
exceptiongroup==1.3.0
executing==2.2.0
fake-useragent==2.2.0
falcon==4.0.2
fastapi==0.115.13
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
fqdn==1.5.1
frozenlist==1.6.0
gast==0.6.0
gdown==5.2.0
git-filter-repo==2.47.0
google-pasta==0.2.0
googlefinance==0.7
graphql-core==3.2.5
greenlet==3.2.3
grpcio==1.73.1
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.37.0
isoduration==20.11.0
jedi==0.19.2
jmespath==1.0.1
joblib==1.5.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema-specifications==2025.4.1
jsonschema==4.25.0
keras==3.10.0
lark==1.2.2
libclang==18.1.1
license-expression==30.4.1
lxml==6.0.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
mongoengine==0.29.1
motor==3.7.1
msgpack==1.1.1
multidict==6.4.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.4.4
numpy==1.26.4
opt_einsum==3.4.0
optree==0.16.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pillow==11.3.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
playwright==1.53.0
pluggy==1.6.0
prometheus_client==0.22.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
py-moneyed==3.0
py-serializable==2.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic-settings==2.9.1
pydantic==2.11.4
pyee==13.0.0
pymongo==4.13.0
pymssql==2.3.4
pyodbc==5.2.0
pyparsing==3.2.3
pystock==0.1.6
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-json-logger==3.3.0
python-multipart==0.0.18
pytz==2025.2
regex==2024.11.6
requests-toolbelt==1.0.0
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.15.3
selenium==4.30.0
sentiment==0.0.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlalchemy==2.0.30
stack-data==0.6.3
starlette==0.46.0
strawberry-graphql[fastapi]>=0.140
#tensorboard-data-server==0.7.2
#tensorboard==2.19.0
#tensorflow-io-gcs-filesystem==0.31.0
#tensorflow==2.19.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio-websocket==0.12.2
trio==0.30.0
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.22.0
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0